+++ src/rr/model/PlantType.java
@@ -1,7 +1,7 @@
 package rr.model;
 
 // Other Imports
-import shared.metadata.Constants;
+import rr.metadata.Constants;
 
 /**
  * The PlantType class extends the SpeciesType class as well as including
