+++ src/rr/model/SpeciesType.java
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.Map;
 
 // Other Imports
-import shared.metadata.Constants;
+import rr.metadata.Constants;
 
 /**
  * The SpeciesType class is an abstract class that is used to represent a single
