+++ src/rr/net/response/ResponseRREndGame.java
@@ -5,8 +5,8 @@
  */
 package rr.net.response;
 
-import shared.metadata.Constants;
-import shared.util.GamePacket;
+import rr.metadata.Constants;
+import rr.util.GamePacket;
 
 /**
  *
